requirements.txt
absl-py==2.1.0
alembic==1.13.3
altair==4.2.2
aniso8601==9.0.1
annotated-types==0.7.0
anyio==4.6.0
appdirs==1.4.4
arrow==1.3.0
astroid==3.3.5
asttokens==2.4.1
astunparse==1.6.3
attrs==24.2.0
backcall==0.2.0
backoff==2.2.1
blinker==1.8.2
boto3==1.35.34
botocore==1.35.34
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
cfgv==3.4.0
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
codecarbon==2.7.1
comm==0.2.2
contourpy==1.3.0
coverage==7.6.3
cryptography==43.0.1
cycler==0.12.1
dacite==1.6.0
dagshub==0.3.38
dagshub-annotation-converter==0.1.1
databricks-sdk==0.33.0
dataclasses-json==0.6.7
debugpy==1.8.6
decorator==5.1.1
Deprecated==1.2.14
dill==0.3.9
distlib==0.3.8
docker==7.1.0
entrypoints==0.4
exceptiongroup==1.2.2
executing==2.1.0
fastapi==0.115.2
filelock==3.16.1
Flask==3.0.3
flatbuffers==24.3.25
fonttools==4.54.1
fusepy==3.0.1
gast==0.4.0
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
gql==3.5.0
graphene==3.3
graphql-core==3.2.4
graphql-relay==3.2.0
great-expectations==1.1.2
greenlet==3.1.1
grpcio==1.66.2
gunicorn==23.0.0
h11==0.14.0
h5py==3.12.1
httpcore==1.0.6
httptools==0.6.4
httpx==0.27.2
identify==2.6.1
idna==3.10
importlib_metadata==8.4.0
iniconfig==2.0.0
ipykernel==6.10.0
ipython==7.9.0
isort==5.13.2
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==7.4.9
jupyter_core==5.7.2
keras==2.10.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.7
libclang==18.1.1
loguru==0.7.2
lxml==5.3.0
Mako==1.3.5
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.22.0
matplotlib==3.9.2
matplotlib-inline==0.1.7
mccabe==0.7.0
mdurl==0.1.2
mistune==3.0.2
ml-dtypes==0.4.1
mlflow==2.16.2
mlflow-skinny==2.16.2
monotonic==1.6
multidict==6.1.0
mypy-extensions==1.0.0
namex==0.0.8
nest-asyncio==1.6.0
nodeenv==1.9.1
numpy==1.26.4
oauthlib==3.2.2
opentelemetry-api==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-semantic-conventions==0.48b0
opt_einsum==3.4.0
optree==0.13.0
packaging==24.1
pandas==2.1.4
parso==0.8.4
pathvalidate==3.2.1
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.4.0
pipdeptree==2.23.4
platformdirs==4.3.6
pluggy==1.5.0
posthog==2.5.0
pre_commit==4.0.0
prometheus_client==0.21.0
prompt-toolkit==2.0.10
protobuf==3.19.6
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyarrow==17.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
pylint==3.3.1
pynvml==11.5.3
pyparsing==3.1.4
pytest==8.3.3
pytest-cov==5.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
questionary==2.0.1
RapidFuzz==3.10.0
referencing==0.35.1
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.9.1
rpds-py==0.20.0
rsa==4.9
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
s3transfer==0.10.2
scikit-learn==1.5.2
scipy==1.14.1
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
SQLAlchemy==2.0.35
sqlparse==0.5.1
stack-data==0.6.3
starlette==0.40.0
tenacity==9.0.0
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.0
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.4.0
threadpoolctl==3.5.0
tomli==2.0.2
tomlkit==0.13.2
toolz==1.0.0
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
treelib==1.7.0
typer==0.12.5
types-python-dateutil==2.9.0.20241003
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
urllib3==2.2.3
uvicorn==0.32.0
uvloop==0.21.0
virtualenv==20.26.6
watchfiles==0.24.0
wcwidth==0.2.13
websockets==13.1
Werkzeug==3.0.4
wrapt==1.16.0
yarl==1.13.1
zipp==3.20.2
